Fix int/long casting for channel ID in broadcast
- We seem to be storing channel IDs as longs in EsBroadcast, but as
ints in ES itself. This means we need to parse them as ints when
reading and also the EsContentTranslator unit tests that an ES
broadcast map need to ensure channel IDs in the map are ints.

+++ atlas-elasticsearch/src/main/java/org/atlasapi/content/EsContentTranslator.java
@@ -461,7 +461,7 @@ public class EsContentTranslator {
         }
 
         return new Broadcast(
-                Id.valueOf((Long) broadcast.get(EsBroadcast.CHANNEL)),
+                Id.valueOf((Integer) broadcast.get(EsBroadcast.CHANNEL)),
                 startDate,
                 endDate,
                 true
